Assessment4-Lakshay Agarwal/requirements.txt
Flask
snowflake-connector-python
gunicorn
